into "Circle" TAB : Create a normal Axis of a Face. New handle of mesh objects. To create a Normal at click location on a Face: - Click first in the view to select and object, - then push the button, - then click on a location on the selected Face. or To create several Normal of the face: - Click first in the view to select and object, - then select one or several points of the face - then push the button. (These selections can also be done into the Combined View)
into "Object" TAB :
Section Sweep:
# Make a loft defined by a list of profiles along a wire.
Will extrude/sweep a Section along a Trajectory like sweep from Part Workbench but:
- the Section center (of Mass) is move at the first point of the Trajectory and;
- the "plane" of the Section is rotate to be perpendicular to the Trajectory.

- Select first one Section wire (Closed wire will generate volumes by default)
(This Section can be a compound from sketch to realize "tube")
- Select one or several wire(s) as Trajectory(ies)
- Then push this button

NB: You can change first:
- Solid option (if toggled will generate a solid for Closed wire Section only) 
- isFrenet option
- All option (means if the trajectory selected is a compound, all sub wires will be used for the sweep)
- Transition Option (Select a Transition option in case of trajectory with several wires; Transition can be:
# 0 (default), 1 (right corners) or 2 (rounded corners).)

into  "Align" TAB :
Align the Face(s) from selected object(s) to the last Face selected.
- Click first to select a Face of an object or several Faces from several objects. 
These objects will be moved.
- Click second to select a Face to align to (the last object is fixed and will never move).
Then Click the button.

NB:
The center of rotation is the center of the bounbing box if possible or 
the center of the Face.
if the Face of the object selected is already aligned to the last one,
a rotation of 180 deg is applied to the object.
In this case the Axis of rotation is Z vector : Base.Vector(0, 0, 1)

Two clicks will rotate by 180 deg the moving objects.

Align the Edge(s) from selected object(s) to the last Edge selected.
- Click first to select an Edge of an object or several Edges from several objects. 
These objects will be moved.
- Click second to select an Edge to align to (the last object is fixed and will never move).
Then Click the button.

NB:
The center of rotation is the center of the bounbing box if possible or 
the center of the Edge.
if the Edge of the object selected is already aligned to the last one,
a rotation of 180 deg is applied to the object.
In this case the Axis of rotation is Z vector : Base.Vector(0, 0, 1)

Two clicks will rotate by 180 deg the moving objects.

Joint Face(s) from selected object(s) to the last Face selected.
- Click first to select a Face of an object or several Faces from several objects. 
These objects will be moved.
- Click second to select a Face to joint to (the last object is fixed and will never move).
Then Click the button.

Two clicks will rotate by 180 deg the moving objects.